backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
alembic==1.14.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.10.6
pydantic-settings==2.8.1
email-validator==2.2.0
reportlab==4.2.5
Pillow==11.1.0
itsdangerous==2.2.0
pytest==8.3.4
httpx==0.28.1
